move entity_endpoint_dict to common #519
