+++ datagateway_api/src/api_start_utils.py
@@ -17,7 +17,7 @@ from datagateway_api.src.resources.entities.entity_endpoint import (
     get_find_one_endpoint,
     get_id_endpoint,
 )
-from datagateway_api.src.resources.entities.entity_endpoint_dict import endpoints
+from datagateway_api.src.common.entity_endpoint_dict import endpoints
 from datagateway_api.src.resources.non_entities.ping_endpoint import ping_endpoint
 from datagateway_api.src.resources.non_entities.sessions_endpoints import (
     session_endpoints,
