+++ datagateway_api/src/common/entity_endpoint_dict.py
